レスポンス全メモリ読み込み関数化
